[feat+fix+refactor]: Prepare 0.2.0-alpha1 with major system monitoring improvements
BREAKING CHANGES:
- Refactored network module to use native macOS APIs instead of command-line utilities
- Modified IOKit trait interface with new SMC key access methods
- Changed memory monitoring interfaces for better async support

Features:
- GPU:
  * Enhanced Apple Silicon hardware detection
  * Added comprehensive GPU characteristics tracking
  * Improved multi-GPU support
  * Better Metal API integration
- Power:
  * New component-level consumption tracking
  * Added async monitoring support
  * Implemented mock testing framework
- Network:
  * Native traffic statistics implementation
  * 64-bit counter support for high bandwidth
  * Automatic fallback mechanisms
- Memory:
  * Added pressure monitoring with thresholds
  * Implemented page fault tracking
  * Enhanced bandwidth monitoring

Fixes:
- IOKit:
  * Resolved critical segfault in service acquisition
  * Improved memory safety in interface implementation
  * Added SMCKeyData_t Default implementation
- Network:
  * Migrated from netstat to sysctlbyname
  * Enhanced traffic statistics reliability
- Memory:
  * Strengthened FFI code safety
  * Improved async monitoring stability
- Process:
  * Enhanced error handling in process tree

Refactor:
- Enhanced memory management in Obj-C interfaces
- Improved error handling across modules
- Reorganised code structure for better maintainability
- Added comprehensive test coverage for new features

Documentation:
- Updated CHANGELOG.md for 0.2.0-alpha1
- Added new examples for power and network monitoring
- Enhanced API documentation for new features

Testing:
- Added GPU hardware detection tests
- Implemented network module native implementation tests
- Created stable mock implementations for power monitoring

+++ src/disk/monitors.rs
@@ -77,10 +77,9 @@ impl DiskHealthMonitor for DiskHealthMonitorImpl {
     }
 
     async fn is_nearly_full(&self) -> Result<bool> {
-        // Default to 90% threshold
-        let total = self.disk.total as f64;
-        let used = (self.disk.total as f64 * 0.9) as f64; // Default to 90% full
-        Ok((used / total) > 0.9)
+        let total = self.disk.total;
+        let used = self.disk.total as f64 * 0.9; // Default to 90% full
+        Ok((used / total as f64) > 0.9)
     }
 
     async fn is_boot_volume(&self) -> Result<bool> {
@@ -194,13 +193,11 @@ impl DiskIOMonitor for DiskIOMonitorImpl {
             Some(last) => {
                 // Simple rate calculation based on the interval
                 let read_rate = ByteSize::from_bytes(
-                    (current.read_bytes.as_bytes() - last.read_bytes.as_bytes())
-                        / self.update_interval.as_secs() as u64,
+                    (current.read_bytes.as_bytes() - last.read_bytes.as_bytes()) / self.update_interval.as_secs(),
                 );
 
                 let write_rate = ByteSize::from_bytes(
-                    (current.write_bytes.as_bytes() - last.write_bytes.as_bytes())
-                        / self.update_interval.as_secs() as u64,
+                    (current.write_bytes.as_bytes() - last.write_bytes.as_bytes()) / self.update_interval.as_secs(),
                 );
 
                 Ok(Transfer {
@@ -215,12 +212,12 @@ impl DiskIOMonitor for DiskIOMonitorImpl {
 
                 let read_rate = ByteSize::from_bytes(
                     (new_current.read_bytes.as_bytes() - current.read_bytes.as_bytes())
-                        / self.update_interval.as_secs() as u64,
+                        / self.update_interval.as_secs(),
                 );
 
                 let write_rate = ByteSize::from_bytes(
                     (new_current.write_bytes.as_bytes() - current.write_bytes.as_bytes())
-                        / self.update_interval.as_secs() as u64,
+                        / self.update_interval.as_secs(),
                 );
 
                 Ok(Transfer {
